// packages/fluentui/src/components/avatar/types.ts
export type FAvatarShape = "circular" | "square";
export type FAvatarActive = "active" | "inactive" | "unset";
export type FAvatarNamedColor =
  | "dark-red"
  | "cranberry"
  | "red"
  | "pumpkin"
  | "peach"
  | "marigold"
  | "gold"
  | "brass"
  | "brown"
  | "forest"
  | "seafoam"
  | "dark-green"
  | "light-teal"
  | "teal"
  | "steel"
  | "blue"
  | "royal-blue"
  | "cornflower"
  | "navy"
  | "lavender"
  | "purple"
  | "grape"
  | "lilac"
  | "pink"
  | "magenta"
  | "plum"
  | "beige"
  | "mink"
  | "platinum"
  | "anchor";
export type FAvatarColor = "neutral" | "brand" | "colorful" | FAvatarNamedColor;